requirements.txt
absl-py==1.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.0
appdirs==1.4.4
async-timeout==4.0.2
attrdict==2.0.1
attrs==23.1.0
audioread==3.0.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.1.1
fastapi==0.96.0
ffmpy==0.3.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
fugashi==1.2.1
google-auth==2.19.0
google-auth-oauthlib==1.0.0
gradio==3.34.0
gradio-client==0.2.6
grpcio==1.54.2
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipadic==1.0.0
jaconv==0.3.4
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lazy-loader==0.2
librosa==0.8.0
linkify-it-py==2.0.2
lit==16.0.5
llvmlite==0.40.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mecab-python3==1.0.6
git+https://github.com/resemble-ai/monotonic_align.git
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
munch==3.0.0
networkx==3.1
numba==0.57.0
numpy==1.23.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
orjson==3.9.0
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
pooch==1.6.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
pydub==0.25.1
Pygments==2.15.1
pykakasi==2.2.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.5
starlette==0.27.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.0
threadpoolctl==3.1.0
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
tqdm==4.65.0
triton==2.0.0
typing-extensions==4.6.2
tzdata==2023.3
uc-micro-py==1.0.2
unidic-lite==1.0.8
urllib3==1.26.16
uvicorn==0.22.0
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
